+++ jaziku/data/maps/shapes/Colombia/Colombia/Colombia.py
@@ -1,9 +1,9 @@
 # Colombia
 
 # Set the four coordinates points for specify grid for region, area or country
-minlat=-5.0
-maxlat=14.0
-minlon=-79.0
-maxlon=-67.0
+minlat=-4.4
+maxlat=13.0
+minlon=-79.2
+maxlon=-66.7
 # set True or False if this region need particular ncl script
 need_particular_ncl_script = True