+++ docs/changelog.rst
@@ -20,6 +20,7 @@ fixes
 
 - Adjust latitude and longitude of Colombia maps
 - restore threshold problem values when run each station
+- Fix path to save correlation maps
 
 
 0.5.0 (**2012-11-22**)