+++ docs/changelog.rst
@@ -21,6 +21,9 @@ fixes
 - Adjust latitude and longitude of Colombia maps
 - restore threshold problem values when run each station
 - Fix path to save correlation maps
+- Fix call mean function in maps
+- Clean some functions
+- Many small adjustments
 
 
 0.5.0 (**2012-11-22**)