+++ docs/changelog.rst
@@ -24,6 +24,7 @@ fixes
 - Fix call mean function in maps
 - Clean some functions
 - Many small adjustments
+- Fix message number of maps process in each grid
 
 
 0.5.0 (**2012-11-22**)